Switch to faster git-prompt plugin
The old git-prompt was slow machine. Found a faster one that does not
use python
https://github.com/woefe/git-prompt.zsh

The change does not help with oh-my-zsh, but that doesn't matter.

The other option would have been to patch oh-my-zsh git-prompt to check
when to run python part, to speed up other dirs, like this
git rev-parse --is-inside-work-tree